+++ src/api/__mock__/section.ts
@@ -1,17 +1,21 @@
-import { SectionResponse } from '../@types/Section';
+import { GetSectionResponse } from '../@types/Section';
 
-export const mockSection: SectionResponse[] = [
+export const mockSection: GetSectionResponse[] = [
   {
-    id: 0,
-    userId: 0,
-    retrospectiveId: 0,
-    sectionContent: 'hee',
+    sectionId: 1,
+    username: 'hope',
+    content: '노션의 페이지 작성을 잘했다.',
+    likeCnt: 3,
+    sectionName: 'Keep',
+    createdDate: '2024-03-31T16:05:20.983Z',
   },
 
   {
-    id: 1,
-    userId: 1,
-    retrospectiveId: 1,
-    sectionContent: 'hee',
+    sectionId: 2,
+    username: 'awesome',
+    content: '노션의 페이지 작성을 잘했다.',
+    likeCnt: 3,
+    sectionName: 'Keep',
+    createdDate: '2024-03-31T16:05:20.983Z',
   },
 ];
