+++ src/pages/WriteRetroReviseTeamPage.tsx
@@ -4,9 +4,9 @@ import { SettingTitle } from '@/components/writeRetro/layout/Setting';
 export const WriteRetroReviseTeamPage = () => {
   return (
     <>
-      <SettingTitle></SettingTitle>
+      <SettingTitle />
       {/* <SettingMenu></SettingMenu> */}
-      <ManageMenu></ManageMenu>
+      <ManageMenu />
     </>
   );
 };
